move moduel_meta to new module module
